% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_proposal.R
\name{submit_proposal}
\alias{submit_proposal}
\title{Submit GSoC Proposal}
\usage{
submit_proposal()
}
\value{
Logical value:
\itemize{
\item TRUE if submitted before the deadline
\item FALSE if submitted after the deadline
}
}
\description{
This function simulates submitting a proposal to Google Summer of Code.
It checks whether the current time is before or after the submission deadline.
}
\details{
The submission deadline is 31st March 2026.
}
\examples{
submit_proposal()

}
